Major rewrites, progress on catcher
Decided to re-re-do some of the base ballplayer classes to make more sense. That threw off some code (and most of it will need to be fixed later) but it's made some progress on the catcher. Human-controlled catcher creates a bug that teleports him to the corner of the screen but where he is controllable in a fixed rectangle (good), and AI catcher makes smooth movements but doesn't move quite far enough (also good)