Progress towards multiple plays
closed #34 The ball re-sets when the ball goes out of bounds or after a goal. Possession changes (or doesn't) depending on the previous play. The AI is comfortable pitching multiple times in a row, but need to iron out AI special pitches, the ball exiting special pitch, and the human player being able to pitch again after the play re-sets