Improved Guard Offense
closed #33 guards will now decide to counterattack if the keeper is the closest player to the ball. That can look like getting open for a shot or trying to get to the midfield. From there, they will be more careful about where they aim, taking proper shots, passing to the ball to the other guard or to the forwards. There isn't usually time to get all the way to the positions they want, but I debugged it where they have to and it works, in case both forwards get injured or something.